Add discovery spec + agent prompt hints to error states
When resources fail registration or discovery, users now see
"Have your agent fix the errors with a prompt or read the discovery spec"
with a copy-to-clipboard prompt and link to /discovery/spec.

- Failed resources are visible by default (no dropdown)
- Covers: registration failed, no discovery found, invalid document,
  skipped resources, partial failures, and per-resource test errors
- Moved DiscoveryActions to shared discovery module to avoid cross-layer imports
- Extracted DiscoveryFixHint component to eliminate duplication
